+++ src/main/java/com/volomak/movieland/service/dto/MovieListDto.java
@@ -6,7 +6,7 @@ public class MovieListDto {
     private Long id;
     private String name;
     private String originalName;
-    private Long year;
+    private int year;
 
     private String description;
     private Double rate;
@@ -37,11 +37,11 @@ public class MovieListDto {
         this.originalName = originalName;
     }
 
-    public Long getYear() {
+    public int getYear() {
         return year;
     }
 
-    public void setYear(Long year) {
+    public void setYear(int year) {
         this.year = year;
     }
